Multi sp harvest
Uneven harvest works with mutiple species, with a distribution of the
cut on the different species depending on a relative abundance of the
species. This distribution is also dependant on an alpha variable that
will force the distribution more on abundant species or less.

I plan to describe this effect in a vignette.

Next step is to include another perturbation function for species.