Delay is a species param and not Forest
Edited delay such it is linked with an IPM and a species and not the
Forest level. This allow to have species with different delay in a
single simulation.

The delay argument in sim_deter_forest is removed and added to the
species function.

A small intro has been added to the vignette and show an increase in
basal area for the species at equilibrium. Need to understand why it
happens.