mv  recFun to ipm and edit sim for 0 BA
Recruitment function is now binded to fit models, IPM and species. This
explain the suppression of the recruit_fun argument when creating
species.

Simulation have been checked with 0 IPM to see if it fully works when
changing the IPM extraction phase of sim. A test with IPM made every 4
BA has also been done to check the same principle.

A bit of doc has been added.