Build matrix on the fly is done
Now we can compute Mu values and store them in correct matrix.

A new function build the IPM matrix corresponding to the BA and climate,
ready to be used at each step of the simulation with some more tweaks.

This function takes 15ms so it's a bit long, and I try to bring time
down.