Fix on def_init_planting #13
Def init planting was not performing as wanted, with lower recruitment
due to an error in the formula.

This function is experimental may be removed since it always draw an
identical distribution, which can then be used by def_init_k()

For even harvest model, the initialisation of the population come after
the time skip, which represent the lag.